Summary: 64-bit fixes
Remove the .in data files.
Make System::get_oparg return intptr_t, not int, so we work on 64-bit systems properly.
Change the String operator[] definition a bit so that we work on more G++ versions.
Fix fileobj.cc in read_token to use ints rather than chars, so we don't crash on OpenPOWER.