- Summary: new string implementation
Landed the new String implementation, which uses just GC'd strings.  This means that, in most cases, we now have a zero-copying, zero-overhead string mechanism.  Unfortunately, there are a few places that still have copying, and a few places we had to add it.  Some of these will be cleaned up shortly, to make them more efficient.