Summary: Scriptix updates
Updates to combat algorithms.  No these aren't even close to what we're going to use long-term.
Reinstated Object::set_name().  Possibly needs some work on the script integration side.
Remove the 'new' keyword from Scriptix, and with it the ability to create new types, and also to create instances of classes.  The former is now unused, and the latter was only used with Arrays.
Created a new array() function in Scriptix to create arrays, similar in spirit to PHP's array() function.